Use \ifthenelse instead of relying on TeX primitives
This fixes the issue reported here:
https://www.mail-archive.com/lyx-users@lists.lyx.org/msg115156.html

The problem is due to the use of the mathpazo package because
it defines the macro \f@family as "ppl\n", i.e., it appends a
newline character at the end. Now, the "\if" TeX primitive
expands its arguments for comparing them and the newline causes
a "missing \begin{document}" error. Instead of complicating the
code to account for this, it is better using the ifthen package.