Fix windres invocation on Windows.
A relatively recent change to windres sources introduced a
regression that breaks many build scripts. Essentially, when
specifying the argument --preprocessor="gcc -E -xc-header -DRC_INVOKED"
this is now taken to literally mean the program with that exact name.

This regression has been reported:
https://sourceware.org/bugzilla/show_bug.cgi?id=27594
but is still not addressed. So, adapt the windres invocation to
the new paradigm.

Patch by Marco Atzeri, slightly modified by me to adapt it to
cross-compilations.