Avoid crash with weird Labeling layout paragraph
For some reason, beginOfBody is set to zero when a paragraph begins with an
environment separator inset or a newline inset.

TextMetrics::labelFill is called in layouts with both labeltype=manual
and margin=manual (basically Labeling layout). It asserts if
beginOfBody is 0 and the first character is not an environment
separator inset. Obviously, the case of a newline inset had been
forgotten.

This is really a corner case, but it has been reported in a real life document.